+++ filebrowser/sites.py
@@ -67,7 +67,7 @@ def get_site_dict(app_name='filebrowser'):
     # Get names of all deployed filebrowser sites with a give app_name
     deployed = get_resolver(get_urlconf()).app_dict[app_name]
     # Get the deployed subset from the cache
-    return dict((k, v) for k, v in _sites_cache[app_name].iteritems() if k in deployed)
+    return dict((k, v) for k, v in _sites_cache[app_name].items() if k in deployed)
 
 
 def register_site(app_name, site_name, site):
@@ -165,7 +165,7 @@ def handle_file_upload(path, file, site):
     try:
         file_path = os.path.join(path, file.name)
         uploadedfile = site.storage.save(file_path, file)
-    except Exception, inst:
+    except Exception as inst:
         raise inst
     return uploadedfile
 
@@ -269,7 +269,7 @@ class FileBrowserSite(object):
         Get all the enabled actions as a list of (name, func). The list
         is sorted alphabetically by actions names
         """
-        res = self._actions.items()
+        res = list(self._actions.items())
         res.sort(key=lambda name_func: name_func[0])
         return res
 
@@ -283,7 +283,7 @@ class FileBrowserSite(object):
         filter_re = []
         for exp in EXCLUDE:
             filter_re.append(re.compile(exp))
-        for k, v in VERSIONS.iteritems():
+        for k, v in VERSIONS.items():
             exp = (r'_%s(%s)$') % (k, '|'.join(EXTENSION_LIST))
             filter_re.append(re.compile(exp, re.IGNORECASE))
 
@@ -372,7 +372,8 @@ class FileBrowserSite(object):
                     messages.add_message(request, messages.SUCCESS, _('The Folder %s was successfully created.') % form.cleaned_data['name'])
                     redirect_url = reverse("filebrowser:fb_browse", current_app=self.name) + query_helper(query, "ot=desc,o=date", "ot,o,filter_type,filter_date,q,p")
                     return HttpResponseRedirect(redirect_url)
-                except OSError, (errno, strerror):
+                except OSError as e:
+                    errno = e.args[0]
                     if errno == 13:
                         form.errors['name'] = forms.util.ErrorList([_('Permission denied.')])
                     else:
@@ -449,7 +450,7 @@ class FileBrowserSite(object):
                 fileobject.delete()
                 signals.filebrowser_post_delete.send(sender=request, path=fileobject.path, name=fileobject.filename, site=self)
                 messages.add_message(request, messages.SUCCESS, _('Successfully deleted %s') % fileobject.filename)
-            except OSError, (errno, strerror):
+            except OSError:
                 # TODO: define error-message
                 pass
         redirect_url = reverse("filebrowser:fb_browse", current_app=self.name) + query_helper(query, "", "filename,filetype")
@@ -493,7 +494,7 @@ class FileBrowserSite(object):
                     else:
                         redirect_url = reverse("filebrowser:fb_browse", current_app=self.name) + query_helper(query, "", "filename")
                     return HttpResponseRedirect(redirect_url)
-                except OSError, (errno, strerror):
+                except OSError:
                     form.errors['name'] = forms.util.ErrorList([_('Error.')])
         else:
             form = ChangeForm(initial={"name": fileobject.filename}, path=path, fileobject=fileobject, filebrowser_site=self)
@@ -537,7 +538,7 @@ class FileBrowserSite(object):
             if len(request.FILES) > 1:
                 return HttpResponseBadRequest('Invalid request! Multiple files included.')
 
-            filedata = request.FILES.values()[0]
+            filedata = list(request.FILES.values())[0]
 
             fb_uploadurl_re = re.compile(r'^.*(%s)' % reverse("filebrowser:fb_upload", current_app=self.name))
             folder = fb_uploadurl_re.sub('', folder)